requirements.txt
python-dotenv
mysql-connector-python
pytest